# Improve ORC double and float performance with batch reads