--- a/presto-orc/src/main/java/io/prestosql/orc/reader/DoubleStreamReader.java
+++ b/presto-orc/src/main/java/io/prestosql/orc/reader/DoubleStreamReader.java
@@ -22,7 +22,6 @@ import io.prestosql.orc.stream.DoubleInputStream;
 import io.prestosql.orc.stream.InputStreamSource;
 import io.prestosql.orc.stream.InputStreamSources;
 import io.prestosql.spi.block.Block;
-import io.prestosql.spi.block.BlockBuilder;
 import io.prestosql.spi.block.RunLengthEncodedBlock;
 import io.prestosql.spi.type.Type;
 import org.openjdk.jol.info.ClassLayout;
@@ -98,7 +97,10 @@ public class DoubleStreamReader
             }
         }
 
-        if (dataStream == null && presentStream != null) {
+        if (dataStream == null) {
+            if (presentStream == null) {
+                throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but data stream is not present");
+            }
             presentStream.skip(nextBatchSize);
             Block nullValueBlock = RunLengthEncodedBlock.create(type, null, nextBatchSize);
             readOffset = 0;
@@ -106,28 +108,28 @@ public class DoubleStreamReader
             return nullValueBlock;
         }
 
-        BlockBuilder builder = type.createBlockBuilder(null, nextBatchSize);
+        Block block;
         if (presentStream == null) {
-            if (dataStream == null) {
-                throw new OrcCorruptionException(streamDescriptor.getOrcDataSourceId(), "Value is not null but data stream is not present");
-            }
-            dataStream.nextVector(type, nextBatchSize, builder);
+            block = dataStream.nextBlock(type, nextBatchSize);
         }
         else {
-            for (int i = 0; i < nextBatchSize; i++) {
-                if (presentStream.nextBit()) {
-                    type.writeDouble(builder, dataStream.next());
-                }
-                else {
-                    builder.appendNull();
-                }
+            boolean[] isNull = new boolean[nextBatchSize];
+            int nullCount = presentStream.getUnsetBits(nextBatchSize, isNull);
+            if (nullCount == 0) {
+                block = dataStream.nextBlock(type, nextBatchSize);
+            }
+            else if (nullCount != nextBatchSize) {
+                block = dataStream.nextBlock(type, isNull);
+            }
+            else {
+                block = RunLengthEncodedBlock.create(type, null, nextBatchSize);
             }
         }
 
         readOffset = 0;
         nextBatchSize = 0;
 
-        return builder.build();
+        return block;
     }
 
     private void openRowGroup()