### Speedup:
```
Benchmark                                  Speedup
BenchmarkStreamReaders.readDoubleNoNull      3.93x
BenchmarkStreamReaders.readDoubleWithNull    1.57x
BenchmarkStreamReaders.readFloatNoNull       5.17x
BenchmarkStreamReaders.readFloatWithNull     1.70x
```

### Before:
```
Benchmark                                  Mode  Cnt  Score   Error  Units
BenchmarkStreamReaders.readDoubleNoNull    avgt   60  0.228 ± 0.004   s/op
BenchmarkStreamReaders.readDoubleWithNull  avgt   60  0.173 ± 0.002   s/op
BenchmarkStreamReaders.readFloatNoNull     avgt   60  0.212 ± 0.003   s/op
BenchmarkStreamReaders.readFloatWithNull   avgt   60  0.179 ± 0.003   s/op
```

### After:
```
Benchmark                                  Mode  Cnt  Score   Error  Units
BenchmarkStreamReaders.readDoubleNoNull    avgt   60  0.058 ± 0.001   s/op
BenchmarkStreamReaders.readDoubleWithNull  avgt   60  0.110 ± 0.002   s/op
BenchmarkStreamReaders.readFloatNoNull     avgt   60  0.041 ± 0.001   s/op
BenchmarkStreamReaders.readFloatWithNull   avgt   60  0.105 ± 0.001   s/op
```